[storage/fio] Use random distribution of filesizes
Previously we were using a single large file for all
tests which will always give poor results with bcache
and is not representative of workloads. This switches
to using varying filesizes between 32m-512m up to a
max of 60g to hopefully simulate filling the bcache
with writeback_percent dirty_data so that we induce
writebacks as part of our write tests.