Add new override check and disable master config checks
This patch fixes how override assertion methods are used such
that we can add arbitratry amounts of them and they will behave
in a consistent way i.e. you can define any number of override
methods on an app and the first to pass will override all the
rest. If no overrides pass on an app then the remaining
methods are run and the overrides are skipped so that they
are not re-run. Only override methods that PASS count towards
the final results.

Also adds a new skip_if_charm_exists override method and applies
it nova-compute vcpu-pin-set config option to skip if sysconfig
charm exists.

Also disables MasterAssertionHelpers for now since they produce
invalid and inconsistent results (see docstring for more info).