SET-1786 Add OpenStack Sunbeam charms to bundle checker
This patch adds OpenStack Sunbeam charms coverage to
the openstack checks. It also adds support for defining
multiple groups of checks within a checks file and splits
the openstack.yaml checks into two groups;
charmed-openstack and sunbeam.

By default the first group (charmed-openstack) is used but
if the type is provided with a group suffix e.g.
-t openstack:sunbeam then the sunbeam group is used.

Also adds aodh charm to charmed-openstack.